# contracts/api/marketplace-central.openapi.yaml
openapi: 3.1.0
info:
  title: Marketplace Central API
  version: 2026-04-02
paths:
...
  /pricing/simulations:
    get:
      summary: List pricing simulations
      operationId: listPricingSimulations
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPricingSimulationsResponse'
    post:
      summary: Run pricing simulation
      operationId: createPricingSimulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunPricingSimulationRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingSimulation'
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /pricing/simulations/batch:
    post:
      summary: Run batch pricing simulation
      operationId: runPricingBatchSimulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunPricingBatchRequest'
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunPricingBatchResponse'
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "503":
          description: Batch simulation unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    PricingSimulation:
      type: object
      required:
        - simulation_id
        - tenant_id
        - product_id
        - account_id
        - margin_amount
        - margin_percent
        - status
      properties:
        simulation_id:
          type: string
        tenant_id:
          type: string
        product_id:
          type: string
        account_id:
          type: string
        margin_amount:
          type: number
          format: double
        margin_percent:
          type: number
          format: double
        status:
          type: string
    BatchSimulationItem:
      type: object
      required:
        - product_id
        - policy_id
        - selling_price
        - cost_amount
        - commission_amount
        - freight_amount
        - fixed_fee_amount
        - margin_amount
        - margin_percent
        - status
        - freight_source
      properties:
        product_id:
          type: string
        policy_id:
          type: string
        selling_price:
          type: number
          format: double
        cost_amount:
          type: number
          format: double
        commission_amount:
          type: number
          format: double
        freight_amount:
          type: number
          format: double
        fixed_fee_amount:
          type: number
          format: double
        margin_amount:
          type: number
          format: double
        margin_percent:
          type: number
          format: double
        status:
          type: string
          enum: [healthy, warning, critical]
        freight_source:
          type: string
          enum: [melhor_envio, fixed, no_dimensions, me_error, me_not_connected]
    RunPricingSimulationRequest:
      type: object
      required:
        - simulation_id
        - product_id
        - account_id
        - base_price_amount
        - commission_percent
        - fixed_fee_amount
        - shipping_amount
        - min_margin_percent
      properties:
        simulation_id:
          type: string
        product_id:
          type: string
        account_id:
          type: string
        base_price_amount:
          type: number
          format: double
        cost_amount:
          type: number
          format: double
          description: "Optional. If not provided, the backend resolves cost from MetalShopping catalog."
        commission_percent:
          type: number
          format: double
        fixed_fee_amount:
          type: number
          format: double
        shipping_amount:
          type: number
          format: double
        min_margin_percent:
          type: number
          format: double
    RunPricingBatchRequest:
      type: object
      required:
        - product_ids
        - policy_ids
        - origin_cep
        - destination_cep
      properties:
        product_ids:
          type: array
          items:
            type: string
        policy_ids:
          type: array
          items:
            type: string
        origin_cep:
          type: string
        destination_cep:
          type: string
        price_source:
          type: string
          enum: [my_price, suggested_price]
        price_overrides:
          type: object
          additionalProperties:
            type: number
            format: double
          description: "Optional. Map of productID::policyID to override price."
    RunPricingBatchResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BatchSimulationItem'
...
    ListPricingSimulationsResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PricingSimulation'
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: >
                Module-specific error code following the pattern MODULE_ENTITY_REASON.
                Known codes include: CATALOG_PRODUCT_NOT_FOUND, CATALOG_ENRICHMENT_INVALID,
                CLASSIFICATIONS_ENTITY_NOT_FOUND, CLASSIFICATIONS_DUPLICATE_NAME,
                MARKETPLACES_ACCOUNT_INVALID, MARKETPLACES_POLICY_INVALID,
                MARKETPLACES_METHOD_NOT_ALLOWED, MARKETPLACES_FEESCHEDULE_PARAM_MISSING,
                MARKETPLACES_DEFINITIONS_FETCH_FAILED, MARKETPLACES_INTERNAL_ERROR,
                PRICING_SIMULATION_INVALID, PRICING_SIMULATION_INVALID_IDENTITY,
                PRICING_SIMULATION_INVALID_BASE_PRICE, PRICING_SIMULATION_INVALID_VALUES,
                PRICING_SIMULATION_NOT_FOUND, PRICING_METHOD_NOT_ALLOWED,
                CONNECTORS_BATCH_NOT_FOUND, CONNECTORS_RETRY_MISSING_PRODUCT,
                INTERNAL_ERROR, INVALID_REQUEST, NOT_FOUND, CONFLICT.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...